Extract normalize.ts from spec.ts
